Nasty hack to update hosts file dynamically
This is really nasty, but saves me from running a separate dnsmasq container.
I have an nginx container running as a proxy for some stuff. To dynamically add the nginx ip to the hosts file (ie to proxy foo.com to the nginx container), I do a quick sed and substitute SUB_ME with the ip address from the ga-dummy container (all brought up together with docker-compose)